Deployment/requirements.txt
streamlit==1.12.0
tensorflow==2.10.0
numpy<1.24
pillow
urllib3
protobuf<3.20
altair<5